requirements.txt
pillow>=9.0.0
numpy>=1.21.0 